# docs/swagger.yaml
definitions:
...
  user.CheckPhoneUserResponse:
    properties:
      userExist:
        description: 用户是否存在
        type: boolean
    type: object
...
  user.LoginRequest:
    properties:
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
    required:
    - password
    - phone
    type: object
  user.LoginResponse:
    properties:
      token:
        description: 鉴权
        type: string
      uid:
        description: 用户ID
        type: string
    type: object
  user.RegisterRequest:
    properties:
      avatar:
        description: 头像
        type: string
      nick_name:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
    required:
    - nick_name
    - password
    - phone
    type: object
  user.SmsLoginRequest:
    properties:
      code:
        description: 验证码
        type: string
      phone:
        description: 手机号
        type: string
    required:
    - code
    - phone
    type: object
  user.SmsLoginResponse:
    properties:
      profileComplete:
        description: 基本资料是否填充完成
        type: boolean
      token:
        description: 验证签名
        type: string
      uid:
        description: 用户ID
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  title: qingwu
  version: "1.0"
paths:
  /admin/user/register:
    post:
      consumes:
      - application/json
      description: 用户添加
      parameters:
      - description: 用户添加参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户添加
      tags:
      - 后台管理-用户
...
  /v1/user/check/exist:
    get:
      description: 检查手机用户是否存在
      parameters:
      - description: 手机号
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.CheckPhoneUserResponse'
      summary: 检查手机用户是否存在
      tags:
      - 用户
...
  /v1/user/login:
    post:
      consumes:
      - application/json
      description: 用户普通登录
      parameters:
      - description: 用户登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
      summary: 用户普通登录
      tags:
      - 用户
  /v1/user/sms/login:
    post:
      consumes:
      - application/json
      description: 用户短信登录
      parameters:
      - description: 参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.SmsLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.SmsLoginResponse'
      summary: 用户短信登录
      tags:
      - 用户
...
swagger: "2.0"